requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2019.9.11        # via requests
cffi==1.12.3              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.7
grpcio-tools==1.23.0
grpcio==1.23.0            # via grpcio-tools
idna==2.8                 # via requests
msgpack==0.6.2
protobuf==3.9.1           # via grpcio-tools
pycparser==2.19           # via cffi
requests==2.22.0
six==1.12.0               # via cryptography, grpcio, protobuf
urllib3==1.25.5           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via protobuf
